+++ packages/dialtone-vue2/components/tab/tab_group.vue
@@ -180,7 +180,7 @@ export default {
     this.updateSelected();
   },
 
-  beforeUpdate () {
+  updated () {
     this.updateSelected();
   },
 
@@ -255,13 +255,11 @@ export default {
     },
 
     getFocusedTabIndex () {
-      // Hot fix https://github.com/dialpad/dialtone/pull/849
-      // The main issue is that this.tabs is not being updated at the time this is being triggered.
-
       const index = this.tabs.findIndex((context) =>
         this.focusId ? context.tabId === `${this.focusId}` : context.isSelected,
       );
 
+      // Safe guard if tab is not found
       return index === -1 ? 0 : index;
     },