+++ packages/dialtone-vue2/localization/index.js
@@ -1,4 +1,4 @@
-import { LocaleManager, RawBundleSource, useI18N } from '@dialpad/i18n-vue2';
+import { LocaleManager, RawBundleSource } from '@dialpad/i18n-vue2';
 
 import enUS from './en-US.ftl?raw';
 import zhCN from './zh-CN.ftl?raw';
@@ -36,6 +36,10 @@ const localeManagerStorageKey = 'user-locale';
  * https://github.com/dialpad/goblin-client-tools/tree/main/packages/i18n#i18n-vue-3-compatible
  */
 export class DialtoneLocalization {
+  /**
+   * @prop {import('@dialpad/i18n').UseI18N} i18n
+   * @private
+   */
   constructor (locale = null) {
     if (typeof DialtoneLocalization.instance === 'object') {
       return DialtoneLocalization.instance;
@@ -66,7 +70,7 @@ export class DialtoneLocalization {
       namespaces: [dialtoneNamespace],
     });
 
-    localeManager.install(dialtoneNamespace);
+    this.i18n = localeManager.useI18N(dialtoneNamespace);
 
     DialtoneLocalization.instance = this;
 
@@ -118,7 +122,7 @@ export class DialtoneLocalization {
    * https://github.com/dialpad/goblin-client-tools/tree/main/packages/i18n#t
    */
   $t (...args) {
-    return useI18N(dialtoneNamespace).$t(...args);
+    return this.i18n.$t(...args);
   }
 
   /**
@@ -128,7 +132,7 @@ export class DialtoneLocalization {
    * https://github.com/dialpad/goblin-client-tools/tree/main/packages/i18n#ta
    */
   $ta (...args) {
-    return useI18N(dialtoneNamespace).$ta(...args);
+    return this.i18n.$ta(...args);
   }
 
   get currentLocale () {
@@ -142,6 +146,6 @@ export class DialtoneLocalization {
     }
 
     this._locale = newLocale;
-    useI18N(dialtoneNamespace).setI18N({ preferredLocale: newLocale }, dialtoneNamespace);
+    this.i18n.setI18N({ preferredLocale: newLocale }, dialtoneNamespace);
   }
 }