+++ packages/dialtone-vue3/components/rich_text_editor/extensions/mentions/MentionComponent.vue
@@ -5,6 +5,7 @@
   >
     <dt-link
       kind="mention"
+      @click.prevent="handleClick"
     >
       {{ text }}
     </dt-link>
@@ -29,6 +30,18 @@ export default {
   computed: {
     text () {
       return '@' + this.$props.node.attrs.name;
+    },
+  },
+
+  methods: {
+    handleClick () {
+      const mentionData = {
+        name: this.$props.node.attrs.name,
+        id: this.$props.node.attrs.id,
+        avatarSrc: this.$props.node.attrs.avatarSrc,
+        contactKey: this.$props.node.attrs.contactKey,
+      };
+      this.$props.editor.emit('mention-click', mentionData);
     },
   },
 };